auth added for journal entry

+++ src/main/java/com/aasrivas/journalApp/controller/JournalEntryController.java
@@ -8,9 +8,12 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/journal")
@@ -22,11 +25,11 @@ public class JournalEntryController {
     @Autowired
     private UserService userService;
 
-    @GetMapping("getAllJournalEntriesOfUser/{userName}")
-    public ResponseEntity<?> getAllJournalEntriesOfUser(@PathVariable String userName) {
+    @GetMapping("getAll")
+    public ResponseEntity<?> getAllJournalEntries() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userName = authentication.getName();
         User user = userService.findByUserName(userName);
-        if (user == null)
-            return new ResponseEntity<>("User not found.", HttpStatus.NOT_FOUND);
         List<JournalEntry> journalEntries = user.getJournalEntries();
         if (journalEntries.isEmpty())
             return new ResponseEntity<>(user.getJournalEntries(), HttpStatus.NOT_FOUND);
@@ -36,44 +39,58 @@ public class JournalEntryController {
 
     @GetMapping("getById/{id}")
     public ResponseEntity<JournalEntry> getJournalEntryById(@PathVariable ObjectId id) {
-        Optional<JournalEntry> byId = journalEntryService.getById(id);
-        if (byId.isPresent())
-            return new ResponseEntity<>(byId.get(), HttpStatus.OK);
-        else
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-
-    }
-
-    @PostMapping("create/{userName}")
-    public ResponseEntity<?> createJournalEntry(@RequestBody JournalEntry entry, @PathVariable String userName) {
-        try {
-            journalEntryService.createEntry(entry, userName);
-            return new ResponseEntity<>(HttpStatus.CREATED);
-        } catch (Exception e) {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userName = authentication.getName();
+        User user = userService.findByUserName(userName);
+        List<JournalEntry> collect = user.getJournalEntries().stream().filter(x -> x.getId().equals(id)).collect(Collectors.toList());
+        if (!collect.isEmpty()) {
+            Optional<JournalEntry> journalEntry = journalEntryService.getById(id);
+            if (journalEntry.isPresent()) {
+                return new ResponseEntity<>(journalEntry.get(), HttpStatus.OK);
+            }
         }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    @PutMapping("user/{userName}/updateById/{id}")
+    @PostMapping("create-entry")
+    public ResponseEntity<?> createJournalEntry(@RequestBody JournalEntry entry) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userName = authentication.getName();
+        journalEntryService.createEntry(entry, userName);
+        return new ResponseEntity<>(HttpStatus.CREATED);
+    }
+
+    @PutMapping("updateById/{id}")
     public ResponseEntity<?> updateJournalEntryById(@RequestBody Map<String, Object> updates,
-                                                    @PathVariable ObjectId id,
-                                                    @PathVariable String userName) {
-        try {
-            journalEntryService.updateById(updates, id, userName);
-            return new ResponseEntity<>(HttpStatus.RESET_CONTENT);
-        } catch (Exception e) {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_MODIFIED);
+                                                    @PathVariable ObjectId id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userName = authentication.getName();
+        User user = userService.findByUserName(userName);
+        List<JournalEntry> collect = user.getJournalEntries().stream().filter(x -> x.getId().equals(id)).collect(Collectors.toList());
+        if (!collect.isEmpty()) {
+            Optional<JournalEntry> journalEntry = journalEntryService.getById(id);
+            if (journalEntry.isPresent()) {
+                journalEntryService.updateById(updates, journalEntry.get());
+                return new ResponseEntity<>(HttpStatus.ACCEPTED);
+            }
         }
+        return new ResponseEntity<>("Entry not found with id: " + id, HttpStatus.NOT_MODIFIED);
     }
 
 
-    @DeleteMapping("user/{userName}/deleteById/{id}")
-    public ResponseEntity<?> deleteJournalEntryById(@PathVariable ObjectId id, @PathVariable String userName) {
-        try {
-            journalEntryService.deleteById(id, userName);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-        } catch (Exception e) {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_ACCEPTABLE);
+    @DeleteMapping("deleteById/{id}")
+    public ResponseEntity<?> deleteJournalEntryById(@PathVariable ObjectId id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userName = authentication.getName();
+        User user = userService.findByUserName(userName);
+        List<JournalEntry> collect = user.getJournalEntries().stream().filter(x -> x.getId().equals(id)).collect(Collectors.toList());
+        if (!collect.isEmpty()) {
+            Optional<JournalEntry> journalEntry = journalEntryService.getById(id);
+            if (journalEntry.isPresent()) {
+                journalEntryService.deleteById(id, userName);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
         }
+        return new ResponseEntity<>("Entry not found with id: " + id, HttpStatus.NOT_ACCEPTABLE);
     }
 }
